+++ sickrage/core/tv/show/__init__.py
@@ -375,18 +375,15 @@ class TVShow(object):
 
     @property
     def is_anime(self):
-        if int(self.anime) > 0:
-            return True
+        return True if int(self.anime) > 0 else False
 
     @property
     def is_sports(self):
-        if int(self.sports) > 0:
-            return True
+        return True if int(self.sports) > 0 else False
 
     @property
     def is_scene(self):
-        if int(self.scene) > 0:
-            return True
+        return True if int(self.scene) > 0 else False
 
     @property
     def network_logo_name(self):